Rename healthService service file
