+++ backend/routes/api.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 const path = require('path');
 const fs = require('fs').promises;
-const { getAllServices, createUserNotificationFromLog } = require('../services/healthService');
+const { getAllServices, createUserNotificationFromLog } = require('../services/fetchServices');
 const serviceControlRouter = require('./serviceControl');
 
 router.use('/service-control', serviceControlRouter);
